+++ code/analyze/main.py
@@ -48,10 +48,10 @@ def get_media_size():
             for line in open(curr_path + file_name, 'r'):
                 curr_media_size = line.strip()
                 if curr_media_size:
-                    total_media_size += int(curr_media_size)
-                    media_size_each_level[curr_level] += int(curr_media_size)
+                    total_media_size += float(curr_media_size)
+                    media_size_each_level[curr_level] += float(curr_media_size)
                     if "insert" in file_name:
-                        insert_all_media_size += int(curr_media_size)
+                        insert_all_media_size += float(curr_media_size)
 
     print("total_media_size: ", total_media_size)
     print("insert_all_media_size: ", insert_all_media_size)
